feat(MOR-911): promote build_hamlib_discovery_payload to public API
Move `build_hamlib_discovery_payload` and its private helpers from the
internal `rigplane.cli._discover_hamlib` module to the public
`rigplane.backends.discovery` module, where `build_setup_discovery_payload`
already lives.  Add a PEP 562 lazy entry in `rigplane/__init__.py` so it is
also importable as `from rigplane import build_hamlib_discovery_payload`.
Keep a re-export shim in `rigplane.cli._discover_hamlib` for backward
compatibility with existing internal callers; no behavior changes.

Canonical new import path:
  from rigplane.backends.discovery import build_hamlib_discovery_payload
  # or
  from rigplane import build_hamlib_discovery_payload

open-core candidate.  Unblocks rigplane-pro from reaching into
`rigplane.cli._discover_hamlib` (follow-up pro PR will drop the
import-linter whitelist).
